moved directory "chess_related" into "environments" made tensorrt-api more roboust to older networks + it expects the following indices now, rather than layer name data [0], value_out [1], policy_out [2] moved file scanning of OpenSpiel files into MODE_OPEN_SPIEL block in CMakeLists.txt fixed tests
