+++ engine/src/environments/chess_related/outputrepresentation.cpp
