+++ engine/src/stateobj.h
@@ -41,8 +41,8 @@ using blaze::DynamicVector;
 #elif MODE_OPEN_SPIEL
 #include "environments/open_spiel/openspielstate.h"
 #else
-#include "chess_related/boardstate.h"
-#include "chess_related/outputrepresentation.h"
+#include "environments/chess_related/boardstate.h"
+#include "environments/chess_related/outputrepresentation.h"
 #endif
 
 #ifdef MODE_POMMERMAN
