+++ engine/tests/tests.h
@@ -32,7 +32,7 @@
 #ifdef BUILD_TESTS
 #include <vector>
 #include <string>
-#include "chess_related/board.h"
+#include "environments/chess_related/board.h"
 
 using namespace std;
